+++ stonks_java/src/main/java/dev/pollito/stonks_java/stock/application/port/in/StockPortIn.java
@@ -11,5 +11,5 @@ public interface StockPortIn {
 
   List<StockPrice> getStocks();
 
-  void setVolatilityMultiplier(double multiplier);
+  void setVolatilityMultiplier(BigDecimal multiplier);
 }
